requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
flask>=2.0.0 
sqlalchemy>=2.0.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
greenlet>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pandas-ta>=0.3.14b0
groq
feedparser>=0.3.0
python-dotenv>=1.0.0
